% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fbins_insta}
\alias{fbins_insta}
\title{Instagram basic insights.
This returns, in a data frame, a summary of Instagram insights within the specified time period. 
Only 30 days' data can be retrieved through the API and this is the full scope of variables.}
\usage{
fbins_insta(start_date, until_date, time_period, page_access_token,
  insta_account)
}
\arguments{
\item{start_date}{The first full day to report, in the format "YYYY-MM-DD" .}

\item{until_date}{The last full day to report, in the format "YYYY-MM-DD" .}

\item{page_access_token}{This must be a valid page access token with sufficient privileges. Visit the Facebook API Graph Explorer to acquire one.}

\item{insta_account}{This is the Instagram business account to be queried.}

\item{period}{One of "day", "week", "days_28", "month", "lifetime", "total_over_range" .}
}
\description{
Instagram basic insights.
This returns, in a data frame, a summary of Instagram insights within the specified time period. 
Only 30 days' data can be retrieved through the API and this is the full scope of variables.
}
\examples{
fbins_insta("start_date", "until_date", "period", "page_access_token", "insta_account")
fbins_insta("2017-01-20", "2017-01-22", "day", "ABCDEFG1234567890ABCDEFG", "1234567890123")
}
\keyword{api}
\keyword{facebook}
\keyword{insights}
\keyword{instagram}
