% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fbins_pxa}
\alias{fbins_pxa}
\title{Facebook Ads Pixel Insights.
This returns, in a data frame, pixel actions such as conversions or adds-to-cart together with a complement of ads stats such as impressions and clicks.
This is applies pagination through the API - so it may take a time to return the results.}
\usage{
fbins_pxa(start_date, until_date, report_level, fb_access_token, account)
}
\arguments{
\item{start_date}{The first full day to report, in the format "YYYY-MM-DD" .}

\item{until_date}{The last full day to report, in the format "YYYY-MM-DD" .}

\item{report_level}{One of "ad", "adset", "campaign" or "account" .}

\item{fb_access_token}{This must be a valid access token with sufficient privileges. Visit the Facebook API Graph Explorer to acquire one.}

\item{account}{This is the ad account, campaign, adset or ad ID to be queried.}
}
\description{
Facebook Ads Pixel Insights.
This returns, in a data frame, pixel actions such as conversions or adds-to-cart together with a complement of ads stats such as impressions and clicks.
This is applies pagination through the API - so it may take a time to return the results.
}
\examples{
fbins_pxa("start_date", "until_date", "report_level", "time_increment", "fb_access_token", "account")
fbins_pxa("2017-01-20", "2017-01-22", "ad", "ABCDEFG1234567890ABCDEFG", "act_12345678")
fbins_pxa("2017-01-20", "2017-01-22", "ad", "ABCDEFG1234567890ABCDEFG", "act_12345678")
}
\keyword{api}
\keyword{facebook}
\keyword{insights}
