% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fbins_summ}
\alias{fbins_summ}
\title{FB summary insights function.
This returns, in a data frame, a summary of FB insights with no breakdown}
\usage{
fbins_summ(start_date, until_date, report_level, time_increment, fb_access_token, fb_ad_account)
}
\arguments{
\item{start_date}{The first full day to report, in the format "YYYY-MM-DD" .
}
\item{until_date}{The last full day to report, in the format "YYYY-MM-DD" .
}
\item{report_level}{One of "ad", "adset", "campaign" or "account" .
}
\item{time_increment}{An integer representing the reporting increment. If blank, defaults to the entire reporting period.
}
\item{fb_access_token}{This must be a valid access token with sufficient privileges. Visit the Facebook API Graph Explorer to acquire one.
}
\item{fb_ad_account}{
This is the Facebook ad account in the format act_xxxxxxx.
}
\description{
FB summary insights function.
This returns, in a data frame, a summary of FB insights with no breakdown
}
\examples{
fbins_summ("start_date", "until_date", "report_level", "time_increment", "fb_access_token", "fb_ad_account")
fbins_summ("2017-01-20", "2017-01-22", "ad", "1", "ABCDEFG1234567890ABCDEFG", "act_123456789"")
fbins_summ("2017-01-20", "2017-01-22", "ad", "", "ABCDEFG1234567890ABCDEFG", "act_123456789")
}
\keyword{api}
\keyword{facebook}
\keyword{insights}
