% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fbins_ag}
\alias{fbins_ag}
\title{FB insights by age & gender function.
This returns all available FB insights per day including age and gender breakdown to the specified report level, and place into a data frame.
When reporting by age and gender the FB API can timeout when reporting on only a few days. This function allows you to pull one day at a time if needed.}
\usage{
fbins_ag(start_date, until_date, report_level, fb_access_token)
}
\arguments{
\item{start_date}{The first full day to report, in the format "YYYY-MM-DD" .}

\item{until_date}{The last full day to report, in the format "YYYY-MM-DD" .}

\item{report_level}{One of "ad", "adset", "campaign" or "account" .}

\item{fb_access_token}{This must be a valid access token with sufficient privileges. Visit the Facebook API Graph Explorer to acquire one.}
}
\description{
FB insights by age & gender function.
This returns all available FB insights per day including age and gender breakdown to the specified report level, and place into a data frame.
When reporting by age and gender the FB API can timeout when reporting on only a few days. This function allows you to pull one day at a time if needed.
}
\examples{
fbins_ag("start_date", "until_date", "report_level", "fb_access_token")
fbins_ag("2017-01-20", "2017-01-22", "ad", "ABCDEFG1234567890ABCDEFG")
}
\keyword{api}
\keyword{facebook}
\keyword{insights}
